Changed line 158 under send()
Changed so that when the Post Request is sent, we get a callback of the object created which is then used to apply the task._id so that a user can delete a task made without needing a refresh. The idea I went off of is that all objects are referenced, so the tag should be applied to the object in memory.